backend/requirements.txt
anthropic
python-dotenv
pytest
pytest-asyncio
httpx
fastapi
uvicorn[standard]
